requirements.txt
torch>=2.0.0
torchvision>=0.15.0
PyYAML>=6.0
hydra-core>=1.3.2
omegaconf>=2.3.0
numpy>=1.22
tqdm>=4.65
pytest>=7.0
